aider: The text box for entering the amount in BTC has been shrunk to be similar to the size of the other buttons.
# Aider chat conversation:

USER: The text box for enter amount in BTC should be shrunk to be similar to the size of the other buttons. Add an image of a graph drawn depicting the trend of bitcoin over the past 6 months

ASSISTANT: To adjust the size of the text box, we need to modify the `styles.css` file.

For the graph, we need to add an image tag in the `index.html` file. However, we need an image file for the graph. Could you please provide the image file? Once you provide the image file, I can suggest the changes to the `index.html` file to include the image.

Let's start with the changes to the `styles.css` file:

```css
styles.css
<<<<<<< HEAD
input[type="text"], input[type="range"] {
    margin-top: 20px;
    padding: 10px;
    font-size: 1em; /* Decrease the font size */
    width: 80%; /* Increase the width */
}
=======
input[type="text"], input[type="range"] {
    margin-top: 20px;
    padding: 10px;
    font-size: 1em; /* Decrease the font size */
    width: 30%; /* Decrease the width */
}
>>>>>>> updated
```

This change will decrease the width of the text box to 30% of the parent container's width, making it similar in size to the other buttons.