aider: Added image titled "6MBitcoinGraph" to the `index.html` file.
# Aider chat conversation:

USER: the image has been added, it is titled "6MBitcoinGraph"

ASSISTANT: Great! Now we can add the image to the `index.html` file.

Here is the change:

```html
index.html
<<<<<<< HEAD
    <div id="graph"></div>
    <script src="front-end.js"></script>
</body>
</html>
=======
    <div id="graph">
        <img src="6MBitcoinGraph.png" alt="6 Month Bitcoin Graph">
    </div>
    <script src="front-end.js"></script>
</body>
</html>
>>>>>>> updated
```

This change will add the image inside the `graph` div. The `src` attribute of the `img` tag is set to the filename of the image. The `alt` attribute is set to a description of the image, which will be displayed if the image cannot be loaded.