aider: Fix "from_str" and "from_id" method errors in main.rs.
# Aider chat conversation:

USER: cargo build --release
   Compiling rgb20-token v0.1.0 (/Users/noahfuery/Documents/CPSCCourses/cpsc370/CPSC-370_BTC_derivates_AI_Chapman/workspaces/rgb/rgb20-token)
error[E0599]: no function or associated item named `from_str` found for struct `ContractId` in the current scope
  --> src/main.rs:75:53
   |
75 |     let contract_id = rgbstd::contract::ContractId::from_str(&bindle.to_string()).unwrap();
   |                                                     ^^^^^^^^ function or associated item not found in `ContractId`
   |
   = help: items from traits can only be used if the trait is in scope
help: the following trait is implemented but not in scope; perhaps add a `use` for it:
   |
1  + use std::str::FromStr;
   |

error[E0599]: no function or associated item named `from_id` found for struct `Consignment` in the current scope
  --> src/main.rs:76:50
   |
76 |     let contract = rgbstd::containers::Contract::from_id(&contract_id).unwrap();
   |                                                  ^^^^^^^ function or associated item not found in `Consignment<false>`

For more information about this error, try `rustc --explain E0599`.
error: could not compile `rgb20-token` (bin "rgb20-token") due to 2 previous errors

ASSISTANT: The errors indicate that the `from_str` method for `ContractId` and the `from_id` method for `Contract` are not found. This is because these methods do not exist for these types in the `rgbstd` crate.

The `ContractId` can be created directly from a `NodeId` which can be created from a string using the `from_hex` method. The `Contract` can be created from a `Bindle` which can be created from a string using the `from_str` method.

Here are the changes that need to be made:

1. Modify the `workspaces/rgb/rgb20-token/src/main.rs` file.

```rust
workspaces/rgb/rgb20-token/src/main.rs
<<<<<<< SEARCH
    let contract_id = rgbstd::contract::ContractId::from_str(&bindle.to_string()).unwrap();
    let contract = rgbstd::containers::Contract::from_id(&contract_id).unwrap();
=======
    let contract_id = rgbstd::contract::ContractId::from(NodeId::from_hex(&bindle.to_string()).unwrap());
    let contract = rgbstd::containers::Contract::from_str(&bindle.to_string()).unwrap();
>>>>>>> REPLACE
```

Please apply these changes and try to compile the project again.